Initialized the hg38 schema as the first revision
Caleydo/tdp_publicdb#64
